Qualify table schema when copying out
Previously this was getting confused because there were two tables named
`products` in different schemas on the default search path. The query that was
constructed was correct for the chosen schema, but the other schema was ahead in
the search path, causing the `COPY (...) TO stdout` query to pick up data from
the wrong table.