Have truncate cascade
I was unable to clone recently due to a FK constraint. I went into the dump and manually added `CASCADE` right before the semicolon, which allowed me to proceed. The comment mentions we can't truncate as we go along, but cascading when we truncate all tables at once should be ok.